Vijaya (#2)
* Create SingleNumberListSol.java

This file contains solution to LeetCode's Single number question's solution. The solution is the basic approach with Time Complexity of O(n2).

* Create SingleNumberHashMapSol.java

This gives solution to Single Number question's solution. The approach used is HashMap which used to keep the numbers in the array as key and their occurence as count.